put hatter helpers in their own module
