+++ src/lib.rs
@@ -2,9 +2,10 @@
 pub mod shell;
 pub mod app;
 pub mod db;
+mod hatter;
 pub mod markdown;
 mod page;
 pub mod sync;
 pub mod utils;
 
-pub use page::Page;
+pub use {crate::hatter::Hatter, page::Page};
